Rename files to avoid camelCase
