+++ src/entity/entity-manager.ts
@@ -6,9 +6,10 @@
  */
 
 import { BitPool } from "@phughesmcr/bitpool";
+
 import { EntityNotFoundError } from "../errors.ts";
-import { isPositiveUint32, isUint32, numberArrayFromString } from "../utils.ts";
 import type { Entity, EntityManagerSerialized } from "../types.ts";
+import { isPositiveUint32, isUint32, numberArrayFromString } from "../utils.ts";
 
 /** An EntityManager is responsible for creating and destroying entities */
 export class EntityManager {
