+++ src/query/query.ts
@@ -5,11 +5,12 @@
  * @license     MIT
  */
 
-import { isObject } from "../utils.ts";
-import { SpecError } from "../errors.ts";
-import { type Component, isValidComponentArray } from "../component/Component.ts";
 import type { BooleanArray } from "@phughesmcr/booleanarray";
+
+import { type Component, isValidComponentArray } from "../component/component.ts";
+import { SpecError } from "../errors.ts";
 import type { QueryInstance, QuerySpec, SchemaOrNull } from "../types.ts";
+import { isObject } from "../utils.ts";
 
 /**
  * Type guard for QuerySpec
