+++ src/types.ts
@@ -5,18 +5,23 @@
  * @license     MIT
  */
 
-import type { Partition, Schema, TypedArray, TypedArrayConstructor } from "@phughesmcr/partitionedbuffer";
-import type { PartitionStorage } from "@phughesmcr/partitionedbuffer";
 import type { BooleanArray } from "@phughesmcr/booleanarray";
+import type {
+  Partition,
+  PartitionStorage,
+  Schema,
+  TypedArray,
+  TypedArrayConstructor,
+} from "@phughesmcr/partitionedbuffer";
 
-import type { Component } from "./component/Component.ts";
-import type { Query } from "./query/Query.ts";
-import type { World } from "./world/World.ts";
-import type { System } from "./system/System.ts";
-import type { ComponentInstance } from "./component/ComponentInstance.ts";
-import type { Archetype } from "./archetype/Archetype.ts";
-import type { StorageProxy } from "./component/StorageProxy.ts";
+import type { Archetype } from "./archetype/archetype.ts";
+import type { ComponentInstance } from "./component/component-instance.ts";
+import type { Component } from "./component/component.ts";
+import type { StorageProxy } from "./component/storage-proxy.ts";
 import type { $_PARTITION_KEY, $_SYSTEM_DESTROY_KEY, $_SYSTEM_INIT_KEY } from "./constants.ts";
+import type { Query } from "./query/query.ts";
+import type { System } from "./system/system.ts";
+import type { World } from "./world/world.ts";
 
 export type { PartitionStorage, Schema, TypedArray, TypedArrayConstructor };
